backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
slowapi==0.1.9

# PDF Generation
reportlab==4.0.7

# Diagram Generation
graphviz==0.20.1

# Input Parsing
PyPDF2==3.0.1
Pillow==10.4.0
pytesseract==0.3.10

# Caching and Task Queue
redis==5.0.1
celery==5.3.4

# Vector Database (ChromaDB for local development)
chromadb==0.4.18

# LLM Integration
openai==1.3.7
anthropic==0.7.7
google-generativeai==0.8.3
httpx==0.25.2  # Pin httpx version for compatibility with openai
# Note: Perplexity uses OpenAI-compatible API, no additional library needed

# Testing
hypothesis==6.92.1
pytest==7.4.3
pytest-asyncio==0.21.1

# Utilities
python-dotenv==1.0.0
pydantic==2.9.2
pydantic-settings==2.6.0

# Authentication & Database
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
alembic==1.13.0
cryptography==41.0.7
